Deleted LoggingConcurrencyMode and pasted code from that to LoggerManager + updated Readme
